app/requirements.txt
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ollama==0.4.7
pydantic==2.10.6
pydantic_core==2.27.2
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
python-dotenv==1.0.1